[error] change double to Double to avoid approximation errors when combining with reals with costs (thanks to Tobias Paxian) ; wcnf reader accepts duplicated literals in the same clause

+++ src/core/tb2knapsack.hpp
@@ -696,7 +696,7 @@ public:
         //        }
         assert(capacity > 0);
         assert(MaxWeight > 0);
-        return ((double)capacity / (double)MaxWeight); // * ucost ???
+        return (double)((Double)capacity / (Double)MaxWeight); // * ucost ???
     }
 
     //TODO: needed for dominance test by DEE
